Checkpoint to be rolled back.
I don't really know the best way to implement the concept of games.

The simple storage that botkit provides is a naive key value store
largely hinged on id (of user, team or channel). Right now, when there's
only one game to track it is easy enough to store each user's elo and
just clobber everything every time but for more games it gets hard.

Plus I want to keep the history of wins etc possibly, which would allow
elos to be recreated and just be neat.

So, I'm thinking it is time for an actual database. I wonder if anyone's
written simple database wrappers for botkit.